refactor Review interfaces
