--- src/frontend/src/components/public/PIAFormTabs/review/interfaces.ts
@@ -1,20 +1,0 @@
-export interface IReview {
-  programArea: {
-    selectedRoles: string[];
-    reviews?: {
-      [role: string]: IReviewSection;
-    };
-  };
-  mpo: IReviewSection;
-  cpo?: {
-    [guid: string]: IReviewSection;
-  };
-}
-
-export interface IReviewSection {
-  isAcknowledged: boolean;
-  reviewNote: string;
-  reviewedByDisplayName?: string;
-  reviewedByGuid?: string;
-  reviewedAt?: string;
-}
