+++ src/frontend/src/utils/statusList/utils.ts
@@ -1,6 +1,6 @@
 import { IPiaForm } from '../../types/interfaces/pia-form.interface';
 import { PiaStatuses, SubmitButtonTextEnum } from '../../constant/constant';
-import { IReviewSection } from '../../components/public/PIAFormTabs/review/interfaces';
+import { IReviewSection } from '../../components/public/PIAFormTabs/review/helpers/interfaces';
 
 export const checkButtonText = (pia: IPiaForm | null) => {
   // in MPO status the button text will different
